* refactored radiusSearchStore into locationSearchStore * renamed SearchBarStore to TextSearchStore
