+++ app/src/components/layout/Header.vue
@@ -98,7 +98,7 @@
 <script lang="ts">
     import Vue from 'vue';
     import { createNamespacedHelpers } from 'vuex';
-    const { mapActions, mapState } = createNamespacedHelpers('searchBarModule');
+    const { mapActions, mapState } = createNamespacedHelpers('textSearchModule');
     import LocationSearchBar from '@/components/ui/LocationSearchBar.vue';
     import Radius from '@/components/ui/Radius.vue';
     import SearchBar from '@/components/ui/SearchBar.vue';
