+++ app/src/store/LocationSearch/actions.ts
@@ -10,4 +10,7 @@ export const actions: ActionTree<LocationSearchState, RootState> = {
     commit('setSelectedLocation', selectedLocation);
     dispatch('findPosts', {}, {root: true});
   },
+  setSelectedRadius({ commit, dispatch }, radiusSearchValue): void {
+    commit('setSelectedRadius', radiusSearchValue);
+  }
 };
