+++ app/src/store/LocationSearch/types.ts
@@ -1,4 +1,5 @@
 export interface LocationSearchState {
   locationSearchValue: string;
   selectedLocation: string;
+  selectedRadius: string;
 }
