--- app/src/store/RadiusSearchStore/actions.ts
@@ -1,9 +1,0 @@
-import {ActionTree} from 'vuex';
-import {RadiusSearchState} from './types';
-import {RootState} from '@/store/types';
-
-export const actions: ActionTree<RadiusSearchState, RootState> = {
-  setSelectedRadius({ commit, dispatch }, radiusSearchValue): void {
-    commit('setSelectedRadius', radiusSearchValue);
-  }
-};
