--- app/src/store/RadiusSearchStore/mutations.ts
@@ -1,8 +1,0 @@
-import { MutationTree } from 'vuex';
-import { RadiusSearchState } from './types';
-
-export const mutations: MutationTree<RadiusSearchState> = {
-  setSelectedRadius(state, value): void {
-    state.selectedRadius = value;
-  },
-};
