--- app/src/store/RadiusSearchStore/types.ts
@@ -1,3 +1,0 @@
-export interface RadiusSearchState {
-  selectedRadius: string;
-}
